Battery Street

+++ 0x02-functions_nested_loops/11-print_to_98.c
@@ -15,7 +15,7 @@ void print_to_98(int n)
 	{
 		while (n < 98)
 		{
-			_putchar(n);
+			_putchar((unsigned int)(n));
 			_putchar(',');
 			n++;
 		}
